refactor: Move files into folder and reorganize
