+++ getrecipe/services/recipe/__init__.py
@@ -1,0 +1,2 @@
+from .main import get_scraped_recipe  # noqa: F401
+from .model import Recipe  # noqa: F401
