+++ getrecipe/services/recipe/model.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel
+from typing import List, Optional
+
+
+class Recipe(BaseModel):
+    url: Optional[str]
+    title: Optional[str]
+    ingredients: Optional[List[str]]
+    instructions: Optional[List[str]]
+    servings: Optional[str]
+    prep_time: Optional[str]
+    cook_time: Optional[str]
